[#155969892] Add elasticsearch retrier (exp backoff) and moved to new package -persistency-.
